% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RgScholar.R
\name{google_Scholar}
\alias{google_Scholar}
\title{Search Google Scholar}
\usage{
google_Scholar(user_query, year_low = NULL, year_high = NULL,
  journal = NULL)
}
\arguments{
\item{user_query}{search term(s) given by user separated by space}

\item{year_low}{Lower limit of the year filter (optional). Default Value = NULL}

\item{year_high}{Upper limit of the year filter (optional).    Default Value = NULL}

\item{journal}{Limit search to a specific journal (optional).    Default value = NULL}
}
\value{
\code{result} dataframe having titles, links and short abstracts of the research papers matching \code{user_query}
}
\description{
Function to search Google Scholar for research papers
}
\examples{
google_Scholar("heart rate") #without filters
google_Scholar("heart rate",year_low = 1989, year_high = 2015,journal = "nature")
}

